mac_nogpu_requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.10.1
aiortc==1.14.0
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.49.0
anyascii==0.3.3
anyio==4.11.0
APScheduler==3.11.1
async-timeout==5.0.1
asyncua==1.1.8
attrs==25.4.0
av==16.0.1
backoff==2.2.1
boto3==1.35.60
botocore==1.35.99
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docopt==0.6.2
easyocr==1.7.2
einops==0.8.1
fastapi==0.115.14
filelock==3.17.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
google-crc32c==1.7.1
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
inference==0.62.0
inference-exp==0.16.3
inference-sdk==0.58.2
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.9
langdetect==1.0.9
lazy_loader==0.4
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.5
ninja==1.13.0
num2words==0.5.14
numpy==2.2.6
nvidia-ml-py==12.575.51
onnx==1.19.1
onnxruntime==1.21.1
onvif-zeep-async==2.0.0
openai==1.109.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.12.0.88
packaging==24.2
paho-mqtt==1.6.1
pandas==2.3.3
peft==0.17.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
polars==1.34.0
polars-runtime-32==1.34.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.23.1
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.0
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pybase64==1.0.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.7.1
pydantic_core==2.33.2
pydot==2.0.0
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==1.0.0
pylogix==1.0.5
pymodbus==3.8.3
PyMsgBox==2.0.1
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pypdfium2==4.30.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-doctr==0.11.0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2025.2
pyvips==2.2.3
PyYAML==6.0.3
qrcode==8.0
RapidFuzz==3.14.3
redis==5.0.8
regex==2025.9.18
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rf-clip==1.1
rich==14.2.0
rubicon-objc==0.5.2
s3transfer==0.10.4
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
SDK==1.0.0
segmentation_models_pytorch==0.5.0
shapely==2.0.7
shellingham==1.5.4
simple-pid==2.0.1
six==1.17.0
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
structlog==24.4.0
supervision==0.27.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.4
timm==1.0.22
tldextract==5.1.3
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.2
twilio==9.3.8
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.217
ultralytics-thop==2.0.17
urllib3==2.5.0
wait_for2==0.3.2
wcwidth==0.2.14
yarl==1.22.0
zeep==4.3.2
zxing-cpp==2.2.0
